posemath: split the header by language
posemath.h declares two libraries in one file: a set of C++ classes with
their operators, and the C API that the real time side uses. The tree already
has a shape for that, in inifile.h and inifile.hh and in hal.h and hal.hh,
where the plain suffix is the C interface and the double one is the C++
interface, so follow it.

posemath.h keeps the C API and gains posemath_types.h, which holds the data
types and the constants. posemath.hh takes the classes, the operators and the
copy templates, and refuses to be compiled as C. posemath.h ends by including
posemath.hh under __cplusplus, so C++ code that includes posemath.h still
gets the classes and nothing out of tree changes. The two headers include
each other and the guards make either order safe.

What this buys is for C++ code: a file that wants no more than PmCartesian,
which is most of what reaches posemath through emcpos.h, can include
posemath_types.h and stop parsing four hundred lines of class declarations.
C compilers were never affected either way, since the classes have always sat
behind an ifdef.

Declaration text is moved verbatim, indentation included. The one edit is
PmAxis, which was an unnamed enum in C and a typedef of the C++ enum PM_AXIS
in C++; the types come first now, so the enum takes its tag and one
definition serves both languages.

Preprocessing posemath.h before and after, in C and in C++, gives the same
declarations in both languages, differing only in the order they appear and
in that PmAxis line. libposemath.so.0 exports the same 390 symbols.
