+++ src/Makefile
@@ -412,6 +412,8 @@ SRCHEADERS := \
     libposemath/gomath.h \
     libposemath/gotypes.h \
     libposemath/posemath.h \
+    libposemath/posemath.hh \
+    libposemath/posemath_types.h \
     libposemath/sincos.h \
     rtapi/rtapi.h \
     rtapi/rtapi_app.h \
